Job-running chosen over non-job running
Servers are checked and the best server with running jobs is chosen and compared to all servers via coreValue.

ie;

A server running a job with a coreScore of 1 will be chosen over a server not running a job also with a coreScore of 1.

A server with a coreScore of 0 that is not running any jobs will be chosen over a server with a coreScore of 1 that is running a job